Move document-related functions to the `documents` module
