+++ src/openprocurement/api/utils/__init__.py
@@ -8,6 +8,7 @@ message = "openprocurement.api.utils now contains submodules. "
 moved('openprocurement.api.utils.api_resource', message)
 moved('openprocurement.api.utils.common', message)
 moved('openprocurement.api.utils.decorators', message)
+moved('openprocurement.api.utils.documents', message)
 moved('openprocurement.api.utils.migration', message)
 moved('openprocurement.api.utils.plugins', message)
 moved('openprocurement.api.utils.searchers', message)
